removed redundant level handler, documented level handler and made it more managable, changed file names
