+++ lib/naer_services/value_utils/handle_enemy_values.dart
